onlybackend

+++ albaing-api/src/main/java/com/jobjob/albaing/dto/JobPost.java
@@ -11,8 +11,8 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class JobPost {
-    private Long jobPostId;                  // 공고 등록 번호
-    private Long companyId;                  // 기업 등록 번호
+    private int jobPostId;                  // 공고 등록 번호
+    private int companyId;                  // 기업 등록 번호
     private String jobPostTitle;             // 공고 제목
     private String jobPostOptionalImage;     // 공고 내용 이미지 주소
     private String jobPostContactNumber;     // 담당자(기업) 연락처
@@ -20,7 +20,7 @@ public class JobPost {
     private String jobPostJobCategory;       // 직군 분류
     private String jobPostJobType;           // 근무 형태
     private String jobPostWorkingPeriod;     // 근무 기간
-    private String jobWorkSchedule;      // 근무 요일
+    private String jobWorkSchedule;          // 근무 요일
     private String jobPostShiftHours;        // 근무 시간
     private String jobPostSalary;            // 급여
     private String jobPostWorkPlace;         // 근무지
@@ -29,4 +29,3 @@ public class JobPost {
     private LocalDateTime jobPostCreatedAt;  // 공고 게시일
     private LocalDateTime jobPostUpdatedAt;  // 공고 최종 수정일
 }
-
