+++ albaing-api/src/main/java/com/jobjob/albaing/mapper/JobPostMapper.java
@@ -9,9 +9,9 @@ import java.util.Map;
 @Mapper
 public interface JobPostMapper {
     void insertJobPost(JobPost jobPost);
-    JobPost selectJobPostById(Long jobPostId);
+    JobPost selectJobPostById(int jobPostId);
     List<JobPost> selectJobPostList(Map<String, Object> params);
     void updateJobPost(JobPost jobPost);
-    void updateJobPostStatus(Long jobPostId, boolean status);
+    void updateJobPostStatus(int jobPostId, boolean status);
     int countJobPost(Map<String, Object> params);
 }
